osquery_kafka_rsyslog/app/requirements.txt
kafka-python==1.4.6
pyyaml==5.1
requests==2.21.0
